CAUSEWAY-3857: standalone PrismHighlighter
