+++ viewers/wicket/ui/src/main/java/org/apache/causeway/viewer/wicket/ui/components/attributes/markup/HighlightBehavior.java
@@ -22,32 +22,21 @@ import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import java.util.concurrent.ConcurrentHashMap;
-import java.util.function.UnaryOperator;
 
 import org.apache.wicket.markup.head.CssHeaderItem;
 import org.apache.wicket.markup.head.IHeaderResponse;
 import org.apache.wicket.request.resource.CssResourceReference;
-import org.graalvm.polyglot.Context;
-import org.jsoup.Jsoup;
-import org.jsoup.nodes.Element;
-import org.jsoup.nodes.Node;
-import org.jsoup.select.NodeVisitor;
 import org.jspecify.annotations.Nullable;
 
 import org.apache.causeway.applib.value.semantics.Renderer.SyntaxHighlighter;
 import org.apache.causeway.commons.internal.exceptions._Exceptions;
-import org.apache.causeway.viewer.commons.prism.PrismLanguage;
-import org.apache.causeway.viewer.commons.prism.PrismNodeHighlighter;
+import org.apache.causeway.viewer.commons.prism.PrismHighlighter;
 import org.apache.causeway.viewer.commons.prism.PrismTheme;
-import org.apache.causeway.viewer.commons.prism.PrismUtils;
 import org.apache.causeway.viewer.wicket.ui.util.PrismResourcesWkt;
-
-import lombok.extern.log4j.Log4j2;
 
 record HighlightBehavior(
     PrismTheme theme,
-    List<CssResourceReference> cssResourceReferences,
-    PrismHighlighter prismHighlighter) {
+    List<CssResourceReference> cssResourceReferences) {
 
     public static Optional<HighlightBehavior> lookup(final @Nullable SyntaxHighlighter syntaxHighlighter) {
         if(syntaxHighlighter==null
@@ -66,7 +55,7 @@ record HighlightBehavior(
     private static final Map<SyntaxHighlighter, HighlightBehavior> cache = new ConcurrentHashMap<>();
     
     private HighlightBehavior(PrismTheme theme) {
-        this(theme, PrismResourcesWkt.cssResources(theme), new PrismHighlighter(PrismUtils.jsResourceMain().orElseThrow()));
+        this(theme, PrismResourcesWkt.cssResources(theme));
     }
         
     void renderHead(final IHeaderResponse response) {
@@ -76,58 +65,8 @@ record HighlightBehavior(
     }
     
     String htmlContentPostProcess(final String htmlContent) {
-        var highlighted = prismHighlighter.apply(htmlContent);
+        var highlighted = new PrismHighlighter().apply(htmlContent);
         return highlighted;
-    }
-    
-    @Log4j2
-    record PrismHighlighter(String prismJs) implements UnaryOperator<String> {
-        
-        /**
-         * Returns the highlighted HTML.
-         * @param htmlContent code to be highlighted
-         */
-        @Override
-        public String apply(String htmlContent) {
-            
-            var doc = Jsoup.parseBodyFragment(htmlContent);
-            
-            var visitor = new NodeVisitor() {
-                
-                @Override
-                public void head(Node node, int depth) {
-                    if(node instanceof Element element
-                        && "code".equals(node.nodeName())) {
-                        
-                        var prismLanguage = PrismLanguage.parseFromCssClass(node.attr("class")).orElse(null);
-                        if(prismLanguage==null) return;
-                        
-                        var grammarJs = PrismUtils.jsResource(prismLanguage).orElse(null);
-                        if(grammarJs==null) {
-                            log.warn("grammarJs not found for {}", prismLanguage);
-                            return;
-                        }
-                        
-                        var newNode = new PrismNodeHighlighter(prismLanguage, ()->{
-                            var context = Context.create("js");
-                            context.eval("js", prismJs);
-                            context.eval("js", grammarJs);
-                            return context;
-                        }).apply(element);
-                        
-                        node.replaceWith(newNode);
-                    }
-                }
-                @Override
-                public void tail(Node node, int depth) {
-                }
-            };
-            
-            doc.traverse(visitor);
-            
-            return doc.body().html();
-        }
-
     }
     
 }
